(WEB-37) Add navbar dropdown for patients
This commit adds a new dropdown to the site navbar which
presents another way to get between the overview dashboard
and patient view.

This is also needed to jump between patients while already in
the patient view of a different patient. Previously, this was
only possible by going back to the overview dashboard and clicking
again on the correct patient.