(WEB-66) Add options menu
I added a dropdown options menu to the overview and patients pages. The overview page allows you to turn on/off both the alerts and the graphs, while the patients page only allows you to turn on/off the graphs. It should pop up an alert notifying you whether it has been turned on/off.