(maint) Move public JS and CSS files into public folder
Previously, our JS and CSS files were scattered around in several
places. It is a better practice to put all of our custom JS and CSS
into a 'public' directory. This will also make it easier for future
maintainers of the project to find necessary code.

In addition, this commit removes our loading of react.js, which
we are not using.

Finally, we had two versions of vertx in our repository, one downloaded
manually and one from bower. The version we were using was the manually
downloaded one, which is newer. This commit replaces bower's version
with that one.